Fix web profile delete functionality
If the web profile was created from different installation then its
profileId was not stored inside configuration table.

Therefore, deletion by profileId failed (or more precisely, it was not
performed).

When we later tried to create profile with the same name, paypal
responsed with error message, because profile already existed.

The fix is to not depend on profileId, but rather on unique profileName.

Closes #38